- Devlog modified. the variable byteswritten used by savetofile() has been redeclared as type long. Now we print the error string when there is a problem with the file descriptor. DATA->size is now of type unsigned long. Now we add a \0 character in every split of piustrsplit()

+++ src/fdhandler/fdhandler.h
@@ -38,12 +38,12 @@
 typedef struct
 {
 	void *data;
-	int size;
+    unsigned long size;
 }DATA;
 
 DATA *loadfile(char * path);
 DATA *loadchkfile(int fd, unsigned long offset, unsigned long size);
-int savetofile(char * path, DATA *data);
+long savetofile(char * path, DATA *data);
 int savetofd(int fd, DATA *data);
 int saveappend(int fd, DATA *data);
 int fileopen(char *path, int create);
